agent_deployment/pyproject.toml
[project]
name = "agent-deployment"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=1.0.5",
    "strands-agents>=1.14.0",
    "strands-agents-tools>=0.2.13",
]
